requirements.txt
pytorch-transformers==1.0.0
numpy==1.16.4
lmdb==0.94
tensorpack==0.9.4
tqdm==4.31.1
easydict==1.9
PyYAML==5.1.2
jsonlines==1.2.0
json-lines==0.5.0
matplotlib
Cython
msgpack
msgpack-numpy